# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Chat de Atendimento - API
  description: |
    API RESTful para sistema de atendimento ao cliente via WhatsApp.
    
    ## Funcionalidades Principais
    
    - 🔐 **Autenticação JWT** com SQL Server
    - 💬 **WhatsApp Business API** (Meta Cloud)
    - 🤖 **Chatbot IA** com Google Gemini
    - ⚡ **WebSocket** para tempo real
    - 📊 **Dashboard** com métricas
    
    ## Fluxo de Autenticação
    
    1. Faça login via `POST /auth/login`
    2. Use o token retornado no header `Authorization: Bearer <token>`
    3. Token expira em 24 horas
    
  version: 2.0.0
  contact:
    name: Suporte API
    email: api@empresa.com
    url: https://empresa.com/suporte
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Ambiente de Desenvolvimento
  - url: https://api.empresa.com/api/v1
    description: Ambiente de Produção
...
paths:
...
  /whatsapp/webhook:
    get:
      tags:
        - whatsapp
      summary: Verificação do webhook
      description: Endpoint de verificação (challenge) do WhatsApp
      operationId: verifyWebhook
      parameters:
        - name: hub.mode
          in: query
          required: true
          schema:
            type: string
        - name: hub.verify_token
          in: query
          required: true
          schema:
            type: string
        - name: hub.challenge
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Challenge aceito
        '403':
          description: Token de verificação inválido

    post:
      tags:
        - whatsapp
      summary: Webhook de eventos
      description: Recebe eventos do WhatsApp Business API
      operationId: handleWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Evento processado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido via /auth/login
...
security:
  - bearerAuth: []